only kill play if a RUNNING_PID exists

+++ restart-all.sh
@@ -5,12 +5,11 @@ export FEDORA_HOME=$ARCHIVE_HOME/fedora
 
 sudo service elasticsearch restart
 
-#cd $ARCHIVE_HOME/regal-api
-#$ARCHIVE_HOME/play-2.2.3/play dist
-#cd -
-
 cd $ARCHIVE_HOME/regal-server
+if [ -f RUNNING_PID ]
+then
 kill `cat RUNNING_PID`
+fi
 nohup $ARCHIVE_HOME/regal-server/bin/regal-api &
 cd -
